tech-job-analyser/data-processing/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0